Fix deskbar replicant
There were various problems which didn't let the deskbar replicant work.
Most notably: the update city message was handled by the window,
which isn't available when in replicant mode.
Moreover, it seems that the URL construction was wrong with certain locales.
Cleanups and simplifications

There is still a problem that the deskbar replicant doesn't handle city changes live: you have to disable
the replicant, close the app,  reenable the deskbar replicant.